Error on extra large messages
Using some large messages on the back end.
Should find a way to reduce messages in the app.
Yet, there was a bug in putting the first message on the queue back on -- this is to do with buffering the message until it  can be fully parsed. (It is well understood that there can be a better buffering mechanism. The one given is a default and there is a config parameter for better implementations)